feat: price: int -> BigDecimal

+++ api/src/main/java/com/github/kastkest/spring_market/api/core/ProductDto.java
@@ -2,6 +2,8 @@ package com.github.kastkest.spring_market.api.core;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
+import java.math.BigDecimal;
+
 @Schema(description = "Модель продукта")
 public class ProductDto {
     @Schema(description = "Id продукта", required = true, example = "1")
@@ -9,12 +11,12 @@ public class ProductDto {
     @Schema(description = "Название продукта", required = true, maxLength = 255, minLength = 3, example = "Milk")
     private String title;
     @Schema(description = "Цена продукта", required = true, example = "200.00")
-    private Integer price;
+    private BigDecimal price;
 
     public ProductDto() {
     }
 
-    public ProductDto(Long id, String title, Integer price) {
+    public ProductDto(Long id, String title, BigDecimal price) {
         this.id = id;
         this.title = title;
         this.price = price;
@@ -36,11 +38,11 @@ public class ProductDto {
         this.title = title;
     }
 
-    public Integer getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 }
